test: split end-to-end tests for mixing prompt_embeds with audio and image embeds in Chat Completion to reduce peak VRAM.
